Remove unnecessary typedef
In C, the type specifier keywords of structures, unions and
enumerations are mandatory.

Using typedef to simplify the declaration of custom type
variables is a common technique. There are opinions that
this is an abuse of typedef, for example Linux kernel coding
style guide:
https://www.kernel.org/doc/html/v4.10/process/coding-style.html#typedefs
tgafunc adopts this suggestion and deletes unnecessary
typedef.