Improve API, modify README.md
Added a function to save TGA image, used to save external image data.
The tga_save() function was renamed to tga_save_from_info(), The new
save function is named tga_save().

Modify some spelling errors in the code.

Update the example code in the README.md according to the adjustment
of the API.